requirements.txt
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
matplotlib==3.8.4
pandas==2.2.2
yagmail==0.15.293